+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `Expression.spread` for the spread operator `...`.
 * Add support 'late' field modifier.
+* Add support for `Expression.nullChecked` to add a null assertion operator.
 
 ## 4.0.0